Revert "Update EffectDatabase.scala"
This reverts commit 862313805f5ab9e1d3be94e06e1e3b38ead013ce.

+++ src/main/scala/com/hoi4utils/clausewitz/code/effect/EffectDatabase.scala
@@ -10,7 +10,6 @@ import com.hoi4utils.clausewitz.map.province.Province
 import com.hoi4utils.clausewitz.map.state.*
 import com.hoi4utils.clausewitz.script.*
 import com.hoi4utils.clausewitz_parser.Node
-import org.apache.logging.log4j.{LogManager, Logger}
 
 import java.io.{File, IOException}
 import java.nio.file.{Files, StandardCopyOption}
@@ -22,8 +21,6 @@ import scala.util.{Try, Using}
 import language.experimental.namedTuples
 
 object EffectDatabase {
-  private val LOGGER: Logger = LogManager.getLogger(getClass)
-
   try {
     Class.forName("org.sqlite.JDBC")
   } catch {
@@ -32,40 +29,29 @@ object EffectDatabase {
   }
 
   private val edb: String = "databases/effects.db"
-  private var _connection: Connection = _ // TODO: @BattleSkorp Replace underscore
+  private var _connection: Connection = _
   private var _effects: List[Effect] = List()
 
   def init(): Unit = {
     try {
-      Option(getClass.getClassLoader.getResource(edb)) match {
-        case Some(url) =>
-          val tempFile = File.createTempFile("effects", ".db")
-          tempFile.deleteOnExit()
+      val url = getClass.getClassLoader.getResource(edb)
+      if (url == null) throw new SQLException(s"Unable to find '$edb'")
 
-          Using(url.openStream) { inputStream =>
-            Files.copy(inputStream, tempFile.toPath, StandardCopyOption.REPLACE_EXISTING)
-          }.recover {
-            case e: Exception =>
-              LOGGER.fatal(s"Failed to copy database file: ${e.getMessage}", e)
-              throw e
-          }
+      val tempFile = File.createTempFile("effects", ".db")
+      tempFile.deleteOnExit()
 
-          if (_connection != null && !_connection.isClosed) {
-            _connection.close()
-          }
-          _connection = DriverManager.getConnection("jdbc:sqlite:" + tempFile.getAbsolutePath)
-          _effects = loadEffects()
-
-        case None =>
-          throw new SQLException(s"Database file '$edb' not found in resources.")
+      Using(url.openStream) { inputStream =>
+        Files.copy(inputStream, tempFile.toPath, StandardCopyOption.REPLACE_EXISTING)
+      }.recover {
+        case e: Exception =>
+          println(s"An error occurred: ${e.getMessage}")
       }
+
+      _connection = DriverManager.getConnection("jdbc:sqlite:" + tempFile.getAbsolutePath)
+      _effects = loadEffects()
     } catch {
-      case e: IOException =>
-        LOGGER.fatal(s"IO error during initialization: ${e.getMessage}", e)
-        throw e
-      case e: SQLException =>
-        LOGGER.fatal(s"SQL error during initialization: ${e.getMessage}", e)
-        throw e
+      case e@(_: IOException | _: SQLException) =>
+        e.printStackTrace()
     }
   }
 
@@ -96,49 +82,57 @@ object EffectDatabase {
       if (_connection != null) _connection.close()
     } catch {
       case e: SQLException =>
-        LOGGER.fatal(s"SQL Error while closing connection: ${e.getMessage}", e)
+        e.printStackTrace()
     }
   }
 
   private def loadEffects(): List[Effect] = {
+    val loadedEffects = new ListBuffer[Effect]
     val retrieveSQL = "SELECT * FROM effects"
+    try {
+      val retrieveStatement = _connection.prepareStatement(retrieveSQL)
+      val resultSet = retrieveStatement.executeQuery
+      while (resultSet.next) {
+        val pdxIdentifier = resultSet.getString("identifier")
+        val supportedScopes_str = resultSet.getString("supported_scopes")
+        val supportedTargets_str = resultSet.getString("supported_targets")
+        val requiredParametersFull_str = resultSet.getString("required_parameters_full")
+        val requiredParametersSimple_str = resultSet.getString("required_parameters_simple")
 
-    Using(_connection.prepareStatement(retrieveSQL)) { retrieveStatement =>
-      Using(retrieveStatement.executeQuery) { resultSet =>
-        val loadedEffects = new ListBuffer[Effect]
+        val supportedScopes = parseEnumSet(supportedScopes_str)
 
-        while (resultSet.next()) {
-          val pdxIdentifier = resultSet.getString("identifier")
-          val supportedScopes = parseEnumSet(resultSet.getString("supported_scopes"))
-          val requiredParametersFull = Option(resultSet.getString("required_parameters_full"))
-          val requiredParametersSimple = Option(resultSet.getString("required_parameters_simple"))
-
+        if (!(requiredParametersFull_str == null && requiredParametersSimple_str == null)) {
           if (supportedScopes.isEmpty) {
-            throw new InvalidParameterException(s"Invalid scope definition for $pdxIdentifier")
+            throw new InvalidParameterException("Invalid scope definition: " + supportedScopes_str)
           }
 
+          val requiredParametersFull = Option(requiredParametersFull_str)
+          val requiredParameterSimple = Option(requiredParametersSimple_str)
+
           val effects = new ListBuffer[Effect]
-          (requiredParametersFull, requiredParametersSimple) match {
-            case (Some(full), Some(simple)) =>
-              effects ++= parametersToEffect(pdxIdentifier, full)
-              effects ++= simpleParameterToEffect(pdxIdentifier, simple)
-            case (Some(full), None) =>
-              effects ++= parametersToEffect(pdxIdentifier, full)
-            case (None, Some(simple)) =>
-              effects ++= simpleParameterToEffect(pdxIdentifier, simple)
-            case (None, None) => ???
+
+          (requiredParametersFull, requiredParameterSimple) match {
+            case (Some(requiredParametersFull), Some(requiredParameterSimple)) =>
+              effects ++= parametersToEffect(pdxIdentifier, requiredParametersFull_str)
+              effects ++= simpleParameterToEffect(pdxIdentifier, requiredParametersSimple_str)
+            case (Some(requiredParametersFull), None) =>
+              effects ++= parametersToEffect(pdxIdentifier, requiredParametersFull_str)
+            case (None, Some(requiredParameterSimple)) =>
+              effects ++= simpleParameterToEffect(pdxIdentifier, requiredParametersSimple_str)
+            case (None, None) =>
+            // todo (bad)
           }
 
           loadedEffects ++= effects
+        } else {
+          System.out.println("No parameters for " + pdxIdentifier + " in effects database.")
         }
-
-        loadedEffects.toList
-      }.getOrElse {
-        throw new SQLException("Failed to execute query or process results.")
       }
-    }.getOrElse {
-      throw new SQLException("Failed to prepare statement.")
+    } catch {
+      case e: SQLException =>
+        e.printStackTrace()
     }
+    loadedEffects.toList
   }
 
   private def parseEnumSet(enumSetString: String): Option[Set[ScopeType]] = {
@@ -344,4 +338,7 @@ object EffectDatabase {
       Some(structuredEffectBlock)
     }
   }
+
+
 }
+
